Fix pegasos kernelz rename;fix default_append_line to not include initrd= as too many arches dont use it by default, test for an initrd in the yaboot config.
git-svn-id: svn+ssh://svn.gentoo.org/var/svnroot/catalyst/trunk@903 d1e1f19c-881f-0410-ab34-b69fee027534

+++ targets/support/functions.sh
@@ -109,7 +109,16 @@ extract_kernels() {
 		rm ${1}/config-* 
 		
 		# change kernel name from "kernel" to "gentoo", for example
-		mv ${1}/kernel-* ${1}/${x}
+		if [ -e ${1}/kernel-* ]
+		then
+		    mv ${1}/kernel-* ${1}/${x}
+		fi
+		
+		# change kernel name from "kernelz" to "gentoo", for example
+		if [ -e ${1}/kernelz-* ]
+		then
+		    mv ${1}/kernelz-* ${1}/${x}
+		fi
 
 		# change initrd name from "initrd" to "gentoo.igz", for example
 		if [ -e ${1}/initrd-* ]
